+++ BKAPI/BK.DAL/ViewModels/Product/VMProductDetails.cs
@@ -34,7 +34,7 @@ public class VMProductDetails
     public int Flute { get; set; }
     public int Back { get; set; }
     public int Ply { get; set; }
-    public int NoOfSheetPerBox { get; set; }
+    public double NoOfSheetPerBox { get; set; }
     public bool IsLamination { get; set; }
     public int? DieCode { get; set; }
     public string PrintTypeName { get; set; }
